requirements.txt
numpy
pandas
scikit-learn
pytest
flake8
requests
fastapi==0.63.0
uvicorn
gunicorn
pyyaml
hydra-core=1.0.6